scripts/requirements.txt
# Dependencies for CSV import script
openai==1.58.1
google-generativeai==0.8.3
supabase==2.10.0
python-dotenv==1.0.1
